Move Averager to used class
